pyproject.toml
[tool.poetry]
name = "rag-with-synthetic-pdf"
version = "0.1.0"
description = ""
authors = ["danielsyahputra <dnlshp@gmail.com>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.63.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
python-docx = "^1.1.2"
docxtpl = "^0.19.1"
docx2pdf = "^0.1.8"
faker = "^36.1.1"
pandas = "^2.2.3"


[tool.poetry.group.notebook.dependencies]
jupyterlab = "^4.3.5"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
